マップ表示処理を別ファイル化
