+++ lib/app_constants.dart
@@ -1,0 +1,14 @@
+class AppConstants {
+  // マップ表示
+  static const double defaultMapLatitude = 35.64691938518296;
+  static const double defaultMapLongitude = 139.71008179999998;
+  static const double defaultMapZoom = 14.0;
+  static const int locationDistanceFilter = 100;
+  static const String shopOpenImagePath = 'assets/shop_open.png';
+  static const String shopCloseImagePath = 'assets/shop_close.png';
+  // 現在地ボタン
+  static const double currentPositionButtonWidth = 55;
+  static const double currentPositionButtonHeight = 55;
+  static const double currentPositionButtonPositionRight = 10;
+  static const double currentPositionButtonPositionBottom = 70;
+}
